chore: Rename store module to catalog
